don't assume the output buffer is initialized
